Refactor main and app modules
