+++ src/lib.rs
@@ -45,6 +45,7 @@ pub mod signals;
 pub mod spotify;
 pub mod station;
 pub mod streaming;
+pub mod terminal_runtime;
 pub mod theme;
 pub mod tools;
 pub mod transfer;
